refactor(github-issues-runtime-tests): extract polling replay module (#1270)
